+++ pkg-manager/headless/src/lockfileToHoistedDepGraph.ts
@@ -40,6 +40,7 @@ export interface LockfileToHoistedDepGraphOptions {
   ignoreScripts: boolean
   currentHoistedLocations?: Record<string, string[]>
   lockfileDir: string
+  modulesDir?: string
   nodeVersion: string
   pnpmVersion: string
   registries: Registries
@@ -83,7 +84,7 @@ async function _lockfileToHoistedDepGraph (
     autoInstallPeers: opts.autoInstallPeers,
   })
   const graph: DependenciesGraph = {}
-  const modulesDir = path.join(opts.lockfileDir, 'node_modules')
+  const modulesDir = path.join(opts.lockfileDir, opts.modulesDir ?? 'node_modules')
   const fetchDepsOpts = {
     ...opts,
     lockfile,